Cleaning up
Separating into multiple files.
Tabs moved, but yet to "un-pub" data;
App tabs changing to refcell to point, instead of indexing;
